refactor: move files
move io functions into io folder
