+++ test/unit/io/packument-io.test.ts
@@ -5,7 +5,7 @@ import { getRegistryPackumentUsing } from "../../../src/io/packument-io";
 import { noopLogger } from "../../../src/logging";
 import { buildPackument } from "../domain/data-packument";
 import { exampleRegistryUrl } from "../domain/data-registry";
-import { mockRegClientGetResult } from "../services/registry-client.mock";
+import { mockRegClientGetResult } from "./registry-client.mock";
 
 describe("packument io", () => {
   describe("fetch", () => {
