+++ test/unit/io/registry-client.mock.ts
